+++ taichi/llvm/llvm_runtime_executor.cpp
@@ -1,7 +1,7 @@
 #include "taichi/llvm/llvm_runtime_executor.h"
 
 #include "taichi/llvm/llvm_offline_cache.h"
-#include "taichi/runtime/llvm/mem_request.h"
+#include "taichi/runtime/llvm/runtime_module/mem_request.h"
 #include "taichi/backends/cpu/cpu_device.h"
 #include "taichi/backends/cuda/cuda_device.h"
 #include "taichi/platform/cuda/detect_cuda.h"
