+++ taichi/runtime/llvm/runtime_module/CMakeLists.txt
@@ -1,4 +1,4 @@
-# ./taichi/runtime/llvm/CMakeLists.txt
+# ./taichi/runtime/llvm/runtime_module/CMakeLists.txt
 
 function(COMPILE_LLVM_RUNTIME rtm_arch)
     message(STATUS "Compiling LLVM byte code file for arch ${rtm_arch}")
@@ -10,7 +10,7 @@ function(COMPILE_LLVM_RUNTIME rtm_arch)
         WORKING_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}"
     )
     add_dependencies(${CORE_LIBRARY_NAME} "generate_llvm_runtime_${rtm_arch}")
-    install(FILES "${CMAKE_SOURCE_DIR}/taichi/runtime/llvm/runtime_${arch}.bc" DESTINATION ${CMAKE_INSTALL_PREFIX}/python/taichi/_lib/runtime)
+    install(FILES "${CMAKE_SOURCE_DIR}/taichi/runtime/llvm/runtime_module/runtime_${arch}.bc" DESTINATION ${CMAKE_INSTALL_PREFIX}/python/taichi/_lib/runtime)
 endfunction()
 
 # Build llvm-runtime for host arch and cuda (if available)
