+++ taichi/system/memory_pool.h
@@ -2,7 +2,7 @@
 #include "taichi/common/core.h"
 #include "taichi/system/unified_allocator.h"
 #define TI_RUNTIME_HOST
-#include "taichi/runtime/llvm/mem_request.h"
+#include "taichi/runtime/llvm/runtime_module/mem_request.h"
 #include "taichi/backends/device.h"
 #include <mutex>
 #include <vector>
